+++ src/App.js
@@ -37,7 +37,6 @@ import { MyOrder } from './pages/mypagelist/myOrder';
 import { MyReviewComment } from './pages/mypagelist/myReviewcomment';
 import { createTheme, ThemeProvider } from "@mui/material";
 import { ApplyListCampaign } from './admin_pages/ApplycampaignList/ApplyCampaignList'
-import { MyProfile } from './pages/mypagelist/myprofile'
 import { MyOrderInfo } from './pages/mypagelist/myorderinfo'
 import { UserWithdrawal } from './pages/UserWithdrawal'
 import { ChatList } from './admin_pages/chatList/ChatList';
@@ -113,7 +112,6 @@ function App() {
             <Route path="/product/:productId" element={<ShopDetail />}></Route>
             <Route path="/product/buy/:productId" element={<BuyProduct />}></Route>
             <Route path="/mypage" element={<Mypage />}></Route>
-            <Route path="/mypage/profile" element={<MyProfile />}></Route>
             <Route path="/mypage/orderinfo" element={<MyOrderInfo />}></Route>
             <Route path="/mypage/profile/withdrawal" element={<UserWithdrawal />}></Route>
             <Route path="/mypage/mypostcampaign" element={<MyPostCampaign />}></Route>
