+++ src/pages/BuyProduct.js
@@ -131,7 +131,7 @@ export default function BuyProduct() {
     }];
 
     try {
-      // await requestPay();
+      await requestPay();
 
 
       const response = await fetch(`${process.env.REACT_APP_BACKEND_URL}/shop/products/order/`, {
