+++ src/pages/mypagelist/myorderinfo.jsx
@@ -96,16 +96,7 @@ export default function MyOrderInfo() {
   const handleFormSubmit = async (e) => {
     e.preventDefault();
     const token = localStorage.getItem("access");
-    // if (
-    //   !userData.address ||
-    //   !userData.zipcode ||
-    //   !userData.detailAddress ||
-    //   !userData.contact ||
-    //   !userData.deliveryMessage
-    // ) {
-    //   alert("모든 필드를 입력해주세요.");
-    //   return;
-    // }
+
     const formData = new FormData();
     formData.append("address", isComplete ? Address : userData.address);
     formData.append("zip_code", isComplete ? zipcode : userData.zipcode);
